+++ program/include/window.hpp
@@ -28,7 +28,7 @@ namespace cse
     void event();
     void input();
     void simulate(const double active_poll_rate);
-    bool start_render(const float aspect_ratio);
+    bool start_render(const double aspect_ratio);
     void end_render();
     void cleanup();
