+++ program/source/camera.cpp
@@ -34,7 +34,7 @@ namespace cse
     hook.call<void(const float)>("simulate", static_cast<float>(active_poll_rate));
   }
 
-  std::pair<glm::mat4, glm::mat4> camera::render(const float aspect_ratio)
+  std::pair<glm::mat4, glm::mat4> camera::render(const double aspect_ratio)
   {
     auto matrices = std::pair{graphics.calculate_projection_matrix(aspect_ratio), state.calculate_view_matrix()};
     hook.call<void(const glm::mat4 &, const glm::mat4 &)>("render", matrices.first, matrices.second);
